source/infrastructure/requirements.txt
click==7.1.2
pytest==6.1.1
pytest-mock==3.2.0
aws_cdk.core==1.70.0
aws_cdk.aws_iam==1.70.0
aws_cdk.aws_lambda==1.70.0
aws_cdk.aws_s3==1.70.0
aws_cdk.aws_s3_deployment==1.70.0
aws_cdk.aws_sns==1.70.0
aws_cdk.aws_stepfunctions==1.70.0
aws_cdk.aws_stepfunctions_tasks==1.70.0
aws_solutions_constructs.aws_lambda_sns==1.70.0
